Добавил круд Тега и политики для него и категорий.

+++ app/Http/Controllers/Api/V1/Content/CategoryController.php
@@ -11,59 +11,83 @@ use App\Http\Resources\CategoryCollection;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
 use App\Services\CategoryService;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Http\Response;
 
 class CategoryController extends Controller
 {
+    use AuthorizesRequests;
+
     public function __construct(protected CategoryService $service)
     {
     }
 
     /**
      * Display a listing of the resource.
+     *
+     * @return CategoryCollection
      */
-    public function index()
+    public function index(): CategoryCollection
     {
+        $this->authorize('viewAny', Category::class);
+
         return new CategoryCollection($this->service->paginate());
     }
 
     /**
      * Store a newly created resource in storage.
+     *
+     * @param StoreCategoryRequest $request
+     * @return CategoryResource
      */
-    public function store(StoreCategoryRequest $request)
+    public function store(StoreCategoryRequest $request): CategoryResource
     {
-        $category = $this->service->create(
-            $request->validated()
-        );
+        $this->authorize('create', Category::class);
+
+        $category = $this->service->create($request->validated());
 
         return new CategoryResource($category);
     }
 
     /**
      * Display the specified resource.
+     *
+     * @param Category $category
+     * @return CategoryResource
      */
-    public function show(Category $category)
+    public function show(Category $category): CategoryResource
     {
+        $this->authorize('view', $category);
+
         return new CategoryResource($category);
     }
 
     /**
      * Update the specified resource in storage.
+     *
+     * @param UpdateCategoryRequest $request
+     * @param Category $category
+     * @return CategoryResource
      */
-    public function update(UpdateCategoryRequest $request, Category $category)
+    public function update(UpdateCategoryRequest $request, Category $category): CategoryResource
     {
-        $updatedCategory = $this->service->update(
-            $category,
-            $request->validated()
-        );
+        $this->authorize('update', $category);
+
+        $updatedCategory = $this->service->update($category, $request->validated());
 
         return new CategoryResource($updatedCategory);
     }
 
     /**
      * Remove the specified resource from storage.
+     *
+     * @param Category $category
+     * @return Response
      */
-    public function destroy(Category $category)
+    public function destroy(Category $category): Response
     {
+        $this->authorize('delete', $category);
+
         $this->service->delete($category);
 
         return response()->noContent();
